Updated cores

+++ flask_s3up/blueprints/route.py
@@ -4,6 +4,7 @@ import os
 
 from werkzeug.wsgi import FileWrapper
 from werkzeug.urls import url_quote
+# from werkzeug.utils import secure_filename
 from flask import Blueprint, Response, request, render_template, current_app
 
 from ..aws.s3 import AWSS3Client
@@ -26,10 +27,13 @@ def get_s3_instance():
         ttl=current_app.config['S3UP_TTL'],
     )
 
-@blueprint.route("/files/<path:key>", methods=['GET', 'DELETE'])
+@blueprint.route("/files/<path:key>", methods=['GET'])
 def files_download(key):
-    # key = urllib.parse.unquote_plus(key)
     if request.method == "GET":
+        """
+        key: encoded
+        """
+        key = urllib.parse.unquote_plus(key)
         s3_client = get_s3_instance()
         r, obj = s3_client.get_object(
             current_app.config['S3UP_BUCKET'],
@@ -59,7 +63,13 @@ def files_download(key):
             rv.headers['Expires'] = '0'
             rv.headers.set('Content-Disposition', 'attachment', **filenames)
             return rv
-    elif request.method == 'DELETE':
+
+@blueprint.route("/files/<path:key>", methods=['DELETE'])
+def files_delete(key):
+    if request.method == 'DELETE':
+        """
+        key: decoded
+        """
         s3_client = get_s3_instance()
         s3_client.delete_objects(
             current_app.config['S3UP_BUCKET'],
@@ -70,17 +80,27 @@ def files_download(key):
 @blueprint.route("/files", methods=['GET', 'POST'])
 def files():
     if request.method == "POST":
+        """
+        prefix: encoded
+        files[].f.filename: decoded
+        prefixer(): 탐색 및 폴더생성시
+        """
+        # form
         prefix = request.form.get('prefix', '')
-        prefix = urllib.parse.unquote(prefix)
+        prefix = urllib.parse.unquote_plus(prefix)
         files = request.files.getlist("files[]")
         s3_client = get_s3_instance()
         prefix = s3_client.prefixer(prefix)
         if not files and prefix:
+            is_exists = s3_client.is_exists(current_app.config['S3UP_BUCKET'], prefix)
+            if is_exists:
+                raise Exception('Already exists')
             s3_client.put_object(
                 current_app.config['S3UP_BUCKET'],
                 prefix,
                 mkdir=True
             )
+            return {}, 201
         else:
             for f in files:
                 f.filename = f'{prefix}{f.filename}'
@@ -89,9 +109,14 @@ def files():
                     f,
                     f.filename
                 )
-        return {}, 201
+            return {}, 201
 
     elif request.method == "GET":
+        """
+        prefix: encoded
+        search: decoded
+        """
+        # args
         prefix = request.args.get('prefix', '')
         prefix = urllib.parse.unquote_plus(prefix)
         starting_token = request.args.get('starting_token')
@@ -123,10 +148,16 @@ def files():
             object_hostname=current_app.config['S3UP_OBJECT_HOSTNAME']
         )
 
+
 @blueprint.context_processor
 def utility_processor():
+    def split(key):
+        return map(lambda k: f'{k}/', key.split('/'))
+
     def unquote_plus(key):
         return urllib.parse.unquote_plus(key)
-    return dict(unquote_plus=unquote_plus)
 
-
+    return dict(
+        split=split,
+        unquote_plus=unquote_plus
+    )
